# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Inventory API",
    "description": "This documentation describes the available endpoints for managing employees and departments.",
    "version": "1.0.0"
  },
  "host": "cse-341-project3-0.onrender.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/employees/": {
      "get": {
        "description": "Get all employees",
        "responses": {
          "200": {
            "description": "A list of employees",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a new employee",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "description": "Get an employee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee found",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "404": {
            "description": "Employee not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "Update an employee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Employee not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Delete an employee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/departments/": {
      "get": {
        "description": "Get all departments",
        "responses": {
          "200": {
            "description": "A list of departments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Department"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a new department",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/departments/{id}": {
      "get": {
        "description": "Get a department by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Department found",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          },
          "404": {
            "description": "Department not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "Update a department by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Department not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Delete a department by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Department deleted successfully"
          },
          "404": {
            "description": "Department not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "departmentId": {
          "type": "string",
          "example": "d001"
        },
        "jobTitle": {
          "type": "string",
          "example": "Software Engineer"
        },
        "salary": {
          "type": "number",
          "example": 75000
        }
      }
    },
    "Department": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Engineering"
        },
        "description": {
          "type": "string",
          "example": "Department for software engineers"
        }
      }
    }
  }
}